refactor: move standalone CLI into TaskiqCMD subcommand
