+++ taskiq/cli/nng/__init__.py
@@ -1,0 +1,1 @@
+"""NNG CLI command."""
